+++ wdom/tests/test_web_node.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-import asyncio
 from unittest.mock import MagicMock, call
 
 from syncer import sync
@@ -207,12 +206,11 @@ class TestQuery(TestCase):
         self.assertEqual(fut.result(), 1)
 
     @sync
-    @asyncio.coroutine
-    def test_scroll(self):
+    async def test_scroll(self):
         fut = self.elm.scrollX()
         self.assertFalse(fut.done())
         self.msg['reqid'] = 0
         self.msg['data'] = {'x': 1}
         self.elm.on_response(self.msg)
-        x = yield from fut
+        x = await fut
         self.assertEqual(x, {'x': 1})